+++ front/app/containers/Admin/dashboard/users/charts/PieChartByCategory.tsx
@@ -6,9 +6,13 @@ import { injectIntl, FormattedMessage } from 'utils/cl-intl';
 import { WrappedComponentProps } from 'react-intl';
 import messages from '../../messages';
 
+// utils
+import renderTooltip from './renderPieChartByCategoryTooltip';
+import { roundPercentages } from 'utils/math';
+
 // styling
 import { withTheme } from 'styled-components';
-import { legacyColors } from 'components/admin/Graphs/styling';
+import { categoricalColorScheme } from 'components/admin/Graphs/styling';
 
 // components
 import ReportExportMenu from 'components/admin/ReportExportMenu';
@@ -28,11 +32,15 @@ import GetSerieFromStream from 'resources/GetSerieFromStream';
 
 // typings
 import { IStreamParams, IStream } from 'utils/streams';
-
+import { LegendItem } from 'components/admin/Graphs/_components/Legend/typings';
 import { IGraphFormat } from 'typings';
 
 interface DataProps {
   serie: IGraphFormat;
+}
+
+interface Serie extends IGraphFormat {
+  percentage: number;
 }
 
 export interface ISupportedDataTypeMap {}
@@ -59,13 +67,6 @@ interface InputProps {
 
 interface Props extends InputProps, DataProps {}
 
-export const piechartColors = [
-  legacyColors.pinkRed,
-  legacyColors.lightBlue,
-  legacyColors.lightGreen,
-  legacyColors.grey,
-];
-
 class PieChartByCategory extends React.PureComponent<
   Props & WrappedComponentProps
 > {
@@ -76,8 +77,22 @@ class PieChartByCategory extends React.PureComponent<
     this.currentChart = React.createRef();
   }
 
+  makeLegends = (row, i): LegendItem => ({
+    icon: 'circle',
+    color: categoricalColorScheme({ rowIndex: i }),
+    label: `${row.name} (${row.percentage}%)`,
+  });
+
+  addPercentages = (serie): Serie | undefined => {
+    if (!serie) return;
+    const percentages = roundPercentages(serie.map((row) => row.value));
+    return serie.map((row, i) => ({
+      ...row,
+      percentage: percentages[i],
+    }));
+  };
+
   render() {
-    const { colorMain } = this.props['theme'];
     const {
       startAt,
       endAt,
@@ -88,6 +103,8 @@ class PieChartByCategory extends React.PureComponent<
       currentGroupFilter,
       currentGroupFilterLabel,
     } = this.props;
+
+    const percentagesSerie = this.addPercentages(serie);
 
     return (
       <GraphCard className={className}>
@@ -106,26 +123,31 @@ class PieChartByCategory extends React.PureComponent<
               />
             )}
           </GraphCardHeader>
-          {!serie ? (
+          {!percentagesSerie ? (
             <NoDataContainer>
               <FormattedMessage {...messages.noData} />
             </NoDataContainer>
           ) : (
             <PieChartStyleFixesDiv>
               <PieChart
-                data={serie}
+                data={percentagesSerie}
+                width={164}
                 mapping={{
                   angle: 'value',
                   name: 'name',
-                  fill: ({ rowIndex }) => piechartColors[rowIndex] ?? colorMain,
                 }}
                 pie={{
                   startAngle: 0,
                   endAngle: 360,
-                  innerRadius: 60,
+                  outerRadius: 60,
                 }}
-                annotations
-                tooltip
+                tooltip={renderTooltip()}
+                legend={{
+                  items: percentagesSerie.map(this.makeLegends),
+                  maintainGraphSize: true,
+                  marginLeft: 50,
+                  position: 'right-center',
+                }}
                 innerRef={this.currentChart}
               />
             </PieChartStyleFixesDiv>